Added answer question action to redux
