+++ src/app/actions.js
@@ -1,7 +1,7 @@
 const Actions = {
   SET_QUESTION_LIST: "SET_QUESTION_LIST",
   SET_CURRENT_EXAM: "SET_CURRENT_EXAM",
-  MARK_QUESTION_ANSWERED: "MARK_QUESTION_ANSWERED",
+  ANSWER_QUESTION: "ANSWER_QUESTION",
 };
 
 export default Actions;
