+++ src/app/store.js
@@ -13,7 +13,7 @@ import examReducer from "./reducers/examReducer";
 
 const store = createStore(
   examReducer,
-  { currentExam: "", questionList: [], answerQuestions: [] },
+  { currentExam: "", questionList: [], answeredQuestions: [] },
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 );
